+++ libsolidity/analysis/ControlFlowGraph.h
@@ -119,7 +119,7 @@ struct FunctionFlow
 	/// (e.g. all return statements of the function).
 	CFGNode* exit = nullptr;
 	/// Revert node. Control flow of the function in case of revert.
-	/// This node is empty does not have any exits, but may have multiple entries
+	/// This node is empty and does not have any exits, but may have multiple entries
 	/// (e.g. all assert, require, revert and throw statements).
 	CFGNode* revert = nullptr;
 	/// Transaction return node. Destination node for inline assembly "return" calls.
@@ -136,6 +136,8 @@ public:
 	bool constructFlow(ASTNode const& _astRoot);
 
 	bool visit(FunctionDefinition const& _function) override;
+	bool visit(ContractDefinition const& _contract) override;
+	void endVisit(ContractDefinition const& _contract) override;
 
 	FunctionFlow const& functionFlow(FunctionDefinition const& _function) const;
 
@@ -155,6 +157,8 @@ private:
 	NodeContainer m_nodeContainer;
 
 	std::map<FunctionDefinition const*, std::unique_ptr<FunctionFlow>> m_functionControlFlow;
+
+	ContractDefinition const* m_contract = nullptr;
 };
 
 }